+++ CHANGELOG.md
@@ -17,6 +17,10 @@
   - Adding sdx-common functionality
   - Updating logger format using sdx-common
   - Adding reprocessing functionality
+  - Updating reprocessing functionality, removing queuepublisher and encrypter
+  - Added Pagination
+  - Added auditable logging
+  - Added searching by valid/invalid
 
 ### 1.0.0 2016-08-16
   - Initial release